[Fabric] Fix BinSkim errors in preparation for 1ES Template migration
This PR fixes the various BinSkim errors detected by running against the
configuration of BinSkim mandated by the 1ES template. This includes
enabling SpectreMitigation, ControlFlowGuard, and CETShadowStack, as
well as removing the warning supressions used by various Fabric-only
files.

Resolves: #12286

+++ vnext/Microsoft.ReactNative/Fabric/Composition/TextInput/WindowsTextInputComponentView.h
@@ -120,7 +120,7 @@ struct WindowsTextInputComponentView : CompositionBaseComponentView {
   std::shared_ptr<facebook::react::WindowsTextInputProps const> m_props;
   std::shared_ptr<facebook::react::WindowsTextInputShadowNode::ConcreteState const> m_state;
   RECT m_rcClient;
-  int m_mostRecentEventCount{0};
+  int64_t m_mostRecentEventCount{0};
   int m_nativeEventCount{0};
   bool m_comingFromJS{false};
   bool m_comingFromState{false};
